Task 1 done
This should complete task 1. The dictionary was modified to use a linked list that keeps track of a word's document frequency. As it reads the file, it splits the information in the file and checks to see if it only contains alphabet chars, and if it does, it keeps track of its frequency. It makes it lowercase. Since the extract list contains stop words, I didn't remove them. It should be easy to do later. I also didn't stem them yet in case they don't want the words stemmed yet.
Right now the linked list only updates itself and displays the document and its frequency. The parts that are needed for part 2 can be added when working in part 2.