keeps track of relative position of text
It keeps track of a word's relative position in text. It does this by keeping track of the order in which the word appears in the document. Stop words do not mess with the position of the text. Ex. "Bob the good" Bob is 0, and good is 1 even though "the" is in between both words.