Part 5 start
Here's what I have for the document correlation list. It takes around 5 mins to do the complete setup for me. I have a function in searcher that should get the top k correlated documents for a doc_id. I didn't test the documents it gets. I just tested to see how long it takes to run. I'll probably test it when I wake up. I didn't change the search loop yet because I'm not sure how you guys want to show the results. This is mostly so that we at least have something.